+++ NEWS.md
@@ -5,6 +5,7 @@
 * Introduce configuration for FSSP (APPPOCTOOL-59)
 * Reinstall endpoint responding with 403 error (MGRENTITLE-134)
 * Validate that upgrade of application does not affect other installed applications (MGRENTITLE-68)
+* Concurrent Kafka topic creation at application flow level causes intermittent errors (MGRENTITLE-135)
 ---
 
 ## Version `v3.1.0` (09.04.2025)